Fixes after code review
1. Refactored code that computes deadline stats and generates priority function in two separate Java classes
2. Refactored core range reading with a generic integer range property reader
