+++ core/src/main/java/org/lflang/generator/c/CGenerator.java
@@ -12,20 +12,14 @@ import static org.lflang.util.StringUtil.addDoubleQuotes;
 
 import com.google.common.base.Objects;
 import com.google.common.collect.Iterables;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import java.io.File;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 import java.util.regex.Pattern;
@@ -36,7 +30,6 @@ import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.xtext.xbase.lib.Exceptions;
 import org.lflang.AttributeUtils;
 import org.lflang.FileConfig;
-import org.lflang.TimeValue;
 import org.lflang.ast.ASTUtils;
 import org.lflang.ast.DelayedConnectionTransformation;
 import org.lflang.federated.extensions.CExtensionUtils;
@@ -655,7 +648,9 @@ public class CGenerator extends GeneratorBase {
                   "    lf_latest_tag_confirmed(tag_to_send);", "    (void) tag_to_send;"),
               "}"));
 
-      generatePriorityFunction();
+      new CPriorityFunctionGenerator(
+              code, targetConfig, main, fileConfig.srcPath, messageReporter, targetLanguageIsCpp())
+          .generate();
 
       // Generate an empty termination function for non-federated
       // execution. For federated execution, an implementation is
@@ -2429,389 +2424,6 @@ public class CGenerator extends GeneratorBase {
   }
 
   /**
-   * Generate a function that provides access to deadline statistics (min, max, median)
-   * of the application. For federated applications, uses federation-level statistics.
-   * For non-federated applications, collects from the main reactor instance.
-   *
-   * This function collects deadlines from all reaction instances in the reactor instance tree,
-   * including those from banks and parameterized reactors, to compute accurate statistics.
-   */
-  private void generatePriorityFunction() {
-    boolean isFederated = targetConfig.isSet(FedSetupProperty.INSTANCE);
-
-    double minDeadline;
-    double maxDeadline;
-    double medianDeadline;
-    List<Double> distinctDeadlinesMs = List.of();
-
-    if (isFederated) {
-      // For federated applications, read deadline stats from the federation properties JSON file
-      Path jsonPath =
-          fileConfig.srcPath.resolve("include" + File.separator + "federation_properties.json");
-
-      try {
-        if (Files.exists(jsonPath)) {
-          String jsonContent = Files.readString(jsonPath, StandardCharsets.UTF_8);
-          JsonObject json = JsonParser.parseString(jsonContent).getAsJsonObject();
-          minDeadline = json.get("minDeadlineMs").getAsDouble();
-          maxDeadline = json.get("maxDeadlineMs").getAsDouble();
-          medianDeadline = json.get("medianDeadlineMs").getAsDouble();
-          if (json.has("deadlinesMs") && json.get("deadlinesMs").isJsonArray()) {
-            JsonArray arr = json.getAsJsonArray("deadlinesMs");
-            List<Double> all = new ArrayList<>(arr.size());
-            for (var el : arr) {
-              if (el != null && el.isJsonPrimitive() && el.getAsJsonPrimitive().isNumber()) {
-                all.add(el.getAsDouble());
-              }
-            }
-            distinctDeadlinesMs = all.stream().distinct().sorted().collect(Collectors.toList());
-          }
-        } else {
-          messageReporter
-              .nowhere()
-              .warning(
-                  "Federation properties JSON file not found at "
-                      + jsonPath
-                      + ", using default values");
-          minDeadline = 0.0;
-          maxDeadline = 0.0;
-          medianDeadline = 0.0;
-        }
-      } catch (IOException | RuntimeException e) {
-        messageReporter
-            .nowhere()
-            .warning(
-                "Failed to read federation properties JSON: "
-                    + e.getMessage()
-                    + ", using default values");
-        minDeadline = 0.0;
-        maxDeadline = 0.0;
-        medianDeadline = 0.0;
-      }
-    } else {
-      // For non-federated applications, collect deadlines from the main reactor instance
-      List<TimeValue> validDeadlines =
-          collectAllDeadlines(main).stream()
-              .filter(
-                  d ->
-                      !TimeValue.NEVER.equals(d)
-                          && !TimeValue.MAX_VALUE.equals(d)
-                          && !TimeValue.FOREVER.equals(d))
-              .sorted()
-              .collect(Collectors.toList());
-
-      if (validDeadlines.isEmpty()) {
-        minDeadline = 0.0;
-        maxDeadline = 0.0;
-        medianDeadline = 0.0;
-      } else {
-        minDeadline = validDeadlines.get(0).toNanoSeconds() / 1000000.0;
-        maxDeadline = validDeadlines.get(validDeadlines.size() - 1).toNanoSeconds() / 1000000.0;
-        int mid = validDeadlines.size() / 2;
-        medianDeadline =
-            (validDeadlines.size() % 2 == 0)
-                ? (validDeadlines.get(mid - 1).toNanoSeconds()
-                        + validDeadlines.get(mid).toNanoSeconds())
-                    / 2
-                    / 1000000.0
-                : validDeadlines.get(mid).toNanoSeconds() / 1000000.0;
-        distinctDeadlinesMs =
-            validDeadlines.stream()
-                .map(d -> d.toNanoSeconds() / 1000000.0)
-                .distinct()
-                .sorted()
-                .collect(Collectors.toList());
-      }
-    }
-
-    // No valid deadlines — all inferred deadlines are the no-deadline sentinel, so every
-    // reaction gets the lowest priority (1), matching the documented no-deadline rule.
-    if (minDeadline == 0.0 && maxDeadline == 0.0 && medianDeadline == 0.0) {
-      prGetPriorityFunction(
-          String.join(
-              "\n",
-              "// Priority assignment function (no deadlines found)",
-              "int get_priority_value(interval_t rel_deadline) {",
-              "    if (rel_deadline <= 0) return 0;",
-              "    return 1;",
-              "}"));
-      return;
-    }
-
-    // Generate the priority function: precompute exp() and derived values in Java so the
-    // generated C uses numeric literals (one exp() remains for the runtime rel_deadline_ms term).
-    if (Double.compare(minDeadline, maxDeadline) == 0) {
-      prGetPriorityFunction(
-          String.join(
-              "\n",
-              "int get_priority_value(interval_t rel_deadline) {",
-              "  if (rel_deadline <= 0) return 0;",
-              // Sentinels encode the absence of a deadline even when every finite deadline
-              // in the program is the same; those reactions get the lowest priority (1).
-              "  if (rel_deadline == FOREVER || rel_deadline >= 281474976710655LL) return 1;",
-              "  return 98;",
-              "}"));
-    } else {
-      // Prefer per-program alpha optimization if we have the full set of deadlines.
-      // (Federated applications get this from federation_properties.json.)
-      final double alpha;
-      if (distinctDeadlinesMs != null && distinctDeadlinesMs.size() >= 2) {
-        alpha = findBestAlphaForDeadlines(distinctDeadlinesMs, minDeadline, maxDeadline);
-      } else {
-        // Fallback: old heuristic (kept only for robustness if deadlinesMs isn't available).
-        final double alphaMax = 0.025;
-        final double alphaMin = 0.005;
-        alpha =
-            alphaMax
-                - (medianDeadline - minDeadline)
-                    / (maxDeadline - minDeadline)
-                    * (alphaMax - alphaMin);
-      }
-      final double expMinusAlphaDmin = Math.exp(-alpha * minDeadline);
-      final double expMinusAlphaDmax = Math.exp(-alpha * maxDeadline);
-      final double denom = expMinusAlphaDmin - expMinusAlphaDmax;
-      final double k = 96.0 / denom;
-      final double p = 98.0 - 96.0 * expMinusAlphaDmin / denom;
-      final double negAlpha = -alpha;
-      // Keep the mapping safe for sentinel "no deadline" values and any values outside the
-      // [minDeadline,maxDeadline] fitting interval.
-      final double maxDeadlineMsFinal = maxDeadline;
-      prGetPriorityFunction(
-          String.join(
-              "\n",
-              "int get_priority_value(interval_t rel_deadline) {",
-              "  if (rel_deadline <= 0) return 0;",
-              // Lingua Franca uses large sentinel values (e.g., TimeValue.MAX_VALUE) to encode
-              // the absence of a deadline. These can be much larger than maxDeadline and would
-              // otherwise map to negative priorities (because g(d) continues decreasing past
-              // d_max).
-              // Treat these as \"no deadline\" -> lowest priority (1).
-              "  if (rel_deadline == FOREVER || rel_deadline >= 281474976710655LL) return 1;",
-              "  double rel_deadline_ms = rel_deadline / 1000000.0;",
-              "  if (rel_deadline_ms >= " + formatDoubleForC(maxDeadlineMsFinal) + ") return 2;",
-              "  const double K = " + formatDoubleForC(k) + ";",
-              "  const double P = " + formatDoubleForC(p) + ";",
-              "  const double neg_alpha = " + formatDoubleForC(negAlpha) + ";",
-              "  double continuous_fun_value = K * exp(neg_alpha * rel_deadline_ms) + P;",
-              "  int prio = (int)round(continuous_fun_value);",
-              "  if (prio < 2) prio = 2;",
-              "  if (prio > 98) prio = 98;",
-              "  return prio;",
-              "}"));
-    }
-  }
-
-  /**
-   * Emit {@code get_priority_value}, which is called from reactor-c (C translation unit).
-   * CCpp compiles the generated main file as C++, so the definition needs C linkage.
-   */
-  private void prGetPriorityFunction(String function) {
-    if (targetLanguageIsCpp()) {
-      code.pr("extern \"C\"");
-    }
-    code.pr(function);
-  }
-
-  /** Smallest alpha (ms^-1) the log-grid search will consider. */
-  private static final double MIN_ALPHA_LIMIT = 1e-12;
-
-  /** Largest alpha (ms^-1) the log-grid search will consider. */
-  private static final double MAX_ALPHA_LIMIT = 1e3;
-
-  /**
-   * Find alpha (ms^-1) that minimizes collisions after rounding, for a given program's distinct
-   * inferred deadlines.
-   *
-   * <p>Optimization goal:
-   * <ol>
-   *   <li>Minimize collision count: {@code collisions = m - |distinct priorities|}.</li>
-   *   <li>Deterministic tie-break for equal collision count: choose smaller alpha.</li>
-   * </ol>
-   *
-   * <p>Note: monotonicity does not need to be enforced explicitly here: the continuous mapping
-   * {@code g(d) = K*exp(-alpha*d)+P} is strictly decreasing in {@code d} for {@code alpha>0}, and
-   * {@code round(·)} is monotone. Therefore integer priorities cannot increase as deadlines increase.
-   */
-  private static double findBestAlphaForDeadlines(
-      List<Double> sortedDistinctDeadlinesMs, double minDeadlineMs, double maxDeadlineMs) {
-    // Search alpha on a log scale: good values can vary over orders of magnitude.
-    //
-    // We start from a tight, canonical bracket derived from two extreme workloads
-    // (see scripts/priority_alpha_search.py), and automatically widen it if the best alpha lands
-    // on a boundary. This keeps compile-time cost low while avoiding missing the optimum.
-    double alphaLo = 1e-5; // ms^-1 (canonical gentle end)
-    double alphaHi = 0.025; // ms^-1 (canonical steep end)
-    final int coarseSteps = 2000;
-    final int fineSteps = 2000;
-
-    // Hard safety limits (only used if boundary widening keeps triggering).
-    final int maxExpansions = 8;
-
-    AlphaSearchOutcome outcome = null;
-    for (int attempt = 0; attempt <= maxExpansions; attempt++) {
-      outcome =
-          searchAlphaWithRefinement(
-              sortedDistinctDeadlinesMs,
-              minDeadlineMs,
-              maxDeadlineMs,
-              alphaLo,
-              alphaHi,
-              coarseSteps,
-              fineSteps);
-      if (outcome == null) {
-        // Every alpha in [alphaLo, alphaHi] underflowed exp(-alpha * d), so the whole
-        // bracket is too steep. Shift the window one decade downward.
-        if (alphaLo > MIN_ALPHA_LIMIT) {
-          alphaHi = alphaLo;
-          alphaLo = Math.max(MIN_ALPHA_LIMIT, alphaLo / 10.0);
-          continue;
-        }
-        return fallbackAlpha(maxDeadlineMs);
-      }
-      if (outcome.hitLowerBoundary && alphaLo > MIN_ALPHA_LIMIT) {
-        alphaLo = Math.max(MIN_ALPHA_LIMIT, alphaLo / 10.0);
-        continue;
-      }
-      if (outcome.hitUpperBoundary && alphaHi < MAX_ALPHA_LIMIT) {
-        alphaHi = Math.min(MAX_ALPHA_LIMIT, alphaHi * 10.0);
-        continue;
-      }
-      break;
-    }
-    return outcome != null ? outcome.best.alpha : fallbackAlpha(maxDeadlineMs);
-  }
-
-  /**
-   * Alpha that keeps {@code exp(-alpha * maxDeadlineMs)} representable in double, used only if
-   * the log-grid search never finds a finite candidate. Do not clamp this up to the search floor
-   * {@code minAlphaLimit}: that floor can still overflow {@code alpha * d} for huge deadlines.
-   */
-  private static double fallbackAlpha(double maxDeadlineMs) {
-    if (!(maxDeadlineMs > 0.0) || !Double.isFinite(maxDeadlineMs)) {
-      return MIN_ALPHA_LIMIT;
-    }
-    // Target alpha * d_max = 1 so neither exponential underflows, even if 1/d_max
-    // is below the search bracket.
-    return 1.0 / maxDeadlineMs;
-  }
-
-  private record AlphaCandidate(double alpha, int collisions) {}
-
-  private record AlphaSearchOutcome(
-      AlphaCandidate best, boolean hitLowerBoundary, boolean hitUpperBoundary) {}
-
-  private record GridSearchResult(AlphaCandidate best, int bestIndex, int steps) {}
-
-  private static AlphaSearchOutcome searchAlphaWithRefinement(
-      List<Double> sortedDistinctDeadlinesMs,
-      double minDeadlineMs,
-      double maxDeadlineMs,
-      double alphaLo,
-      double alphaHi,
-      int coarseSteps,
-      int fineSteps) {
-    final GridSearchResult coarse =
-        bestAlphaOnLogGrid(
-            sortedDistinctDeadlinesMs, minDeadlineMs, maxDeadlineMs, alphaLo, alphaHi, coarseSteps);
-    if (coarse == null) return null;
-
-    // Refine around the coarse winner (in log space).
-    final double lo = Math.log10(alphaLo);
-    final double hi = Math.log10(alphaHi);
-    final double step = (hi - lo) / (coarseSteps - 1);
-    final double center = Math.log10(coarse.best.alpha);
-    final double refineLo = Math.max(lo, center - 3.0 * step);
-    final double refineHi = Math.min(hi, center + 3.0 * step);
-    final GridSearchResult fine =
-        bestAlphaOnLogGrid(
-            sortedDistinctDeadlinesMs,
-            minDeadlineMs,
-            maxDeadlineMs,
-            Math.pow(10.0, refineLo),
-            Math.pow(10.0, refineHi),
-            fineSteps);
-    if (fine == null) return null;
-
-    final boolean hitLower = fine.bestIndex == 0;
-    final boolean hitUpper = fine.bestIndex == fine.steps - 1;
-    return new AlphaSearchOutcome(fine.best, hitLower, hitUpper);
-  }
-
-  private static GridSearchResult bestAlphaOnLogGrid(
-      List<Double> sortedDistinctDeadlinesMs,
-      double minDeadlineMs,
-      double maxDeadlineMs,
-      double alphaLo,
-      double alphaHi,
-      int steps) {
-    AlphaCandidate best = null;
-    int bestIndex = -1;
-    final double lo = Math.log10(alphaLo);
-    final double hi = Math.log10(alphaHi);
-    for (int i = 0; i < steps; i++) {
-      final double t = (double) i / (steps - 1);
-      final double alpha = Math.pow(10.0, lo + t * (hi - lo));
-      final AlphaCandidate cand =
-          scoreAlpha(alpha, sortedDistinctDeadlinesMs, minDeadlineMs, maxDeadlineMs);
-      if (cand == null) continue;
-      if (best == null) {
-        best = cand;
-        bestIndex = i;
-        continue;
-      }
-      if (cand.collisions < best.collisions) {
-        best = cand;
-        bestIndex = i;
-      } else if (cand.collisions == best.collisions && cand.alpha < best.alpha) {
-        best = cand;
-        bestIndex = i;
-      }
-    }
-    return best != null ? new GridSearchResult(best, bestIndex, steps) : null;
-  }
-
-  /**
-   * Score a candidate alpha on the program deadlines.
-   */
-  private static AlphaCandidate scoreAlpha(
-      double alpha,
-      List<Double> sortedDistinctDeadlinesMs,
-      double minDeadlineMs,
-      double maxDeadlineMs) {
-    if (!(alpha > 0.0) || !Double.isFinite(alpha)) return null;
-    if (sortedDistinctDeadlinesMs == null || sortedDistinctDeadlinesMs.size() < 2) return null;
-
-    final double expMinusAlphaDmin = Math.exp(-alpha * minDeadlineMs);
-    final double expMinusAlphaDmax = Math.exp(-alpha * maxDeadlineMs);
-    final double denom = expMinusAlphaDmin - expMinusAlphaDmax;
-    if (denom == 0.0 || !Double.isFinite(denom)) return null;
-
-    final double K = 96.0 / denom;
-    final double P = 98.0 - 96.0 * expMinusAlphaDmin / denom;
-
-    final Set<Integer> used = new HashSet<>();
-    for (double d : sortedDistinctDeadlinesMs) {
-      final double g = K * Math.exp(-alpha * d) + P;
-      final int p = (int) Math.round(g);
-      used.add(p);
-    }
-
-    final int m = sortedDistinctDeadlinesMs.size();
-    final int distinctPriorities = used.size();
-    final int collisions = m - distinctPriorities;
-    return new AlphaCandidate(alpha, collisions);
-  }
-
-  /** Format a finite double as a C floating literal (decimal, US locale). */
-  private static String formatDoubleForC(double value) {
-    if (!Double.isFinite(value)) {
-      throw new IllegalArgumentException(
-          "Non-finite double cannot be emitted as C literal: " + value);
-    }
-    return String.format(Locale.US, "%.17g", value);
-  }
-
-  /**
    * Map a policy name string (from the @platform attribute's scheduler parameter) to the
    * corresponding C #define value.
    *
@@ -2825,30 +2437,5 @@ public class CGenerator extends GeneratorBase {
       case "normal" -> "LF_SCHED_FAIR";
       default -> null;
     };
-  }
-
-  /**
-   * Recursively collect all inferred deadlines from a ReactorInstance.
-   *
-   * <p>This uses inferred deadlines which include deadline propagation through the reaction graph —
-   * if a downstream reaction has an earlier deadline, it is propagated to upstream reactions.
-   *
-   * @param instance The reactor instance to collect deadlines from.
-   * @return A list of all deadlines found in this instance and its children.
-   */
-  private List<TimeValue> collectAllDeadlines(ReactorInstance instance) {
-    List<TimeValue> deadlines = new ArrayList<>();
-
-    // Collect inferred deadlines from all reactions in this instance
-    for (ReactionInstance reaction : instance.reactions) {
-      deadlines.addAll(reaction.getInferredDeadlinesList());
-    }
-
-    // Recursively collect deadlines from child reactor instances
-    for (ReactorInstance child : instance.children) {
-      deadlines.addAll(collectAllDeadlines(child));
-    }
-
-    return deadlines;
   }
 }
